fix(webmcp): dispatch execute_command through the recorded live path
